+++ tests/Console/FileControllerTest.php
@@ -11,6 +11,7 @@ use Hirtz\Media\Test\Fixtures\FolderFixture;
 use Hirtz\Media\Test\Models\TestAsset;
 use Hirtz\Media\Test\Models\TestAssetModel;
 use Hirtz\Media\Test\TestCase;
+use Hirtz\Media\Test\Traits\MediaFileTrait;
 use Hirtz\Skeleton\Helpers\FileHelper;
 use Hirtz\Skeleton\Test\Traits\StdOutBufferControllerTrait;
 use Override;
@@ -21,7 +22,7 @@ use Yii;
  */
 class FileControllerTest extends TestCase
 {
-    private Folder $folder;
+    use MediaFileTrait;
 
     /**
      * @return array<string, mixed>
@@ -94,28 +95,6 @@ class FileControllerTest extends TestCase
         self::assertTrue($asset->insert(), print_r($asset->getErrors(), true));
 
         return $asset;
-    }
-
-    private function createFile(string $basename): File
-    {
-        $path = $this->folder->getUploadPath() . "$basename.jpg";
-
-        $image = imagecreatetruecolor(100, 100);
-        imagejpeg($image, $path);
-
-        $file = File::create();
-        $file->loadDefaultValues();
-        $file->name = ucfirst($basename);
-        $file->basename = $basename;
-        $file->extension = 'jpg';
-        $file->width = 100;
-        $file->height = 100;
-        $file->size = filesize($path) ?: 0;
-        $file->populateFolderRelation($this->folder);
-
-        self::assertTrue($file->insert(), print_r($file->getErrors(), true));
-
-        return $file;
     }
 }