+++ tests/Models/Actions/SaveFolderRedirectsTest.php
@@ -10,6 +10,7 @@ use Hirtz\Media\Models\Folder;
 use Hirtz\Media\Test\Fixtures\FileFixture;
 use Hirtz\Media\Test\Fixtures\FolderFixture;
 use Hirtz\Media\Test\TestCase;
+use Hirtz\Media\Test\Traits\MediaFileTrait;
 use Hirtz\Skeleton\Helpers\FileHelper;
 use Hirtz\Skeleton\Models\Redirect;
 use Override;
@@ -20,7 +21,7 @@ use Override;
  */
 class SaveFolderRedirectsTest extends TestCase
 {
-    private Folder $folder;
+    use MediaFileTrait;
 
     /**
      * @return array<string, mixed>
@@ -196,26 +197,4 @@ class SaveFolderRedirectsTest extends TestCase
         return (int)Redirect::find()->count();
     }
 
-    private function createFile(string $basename): File
-    {
-        $path = $this->folder->getUploadPath() . "$basename.jpg";
-        FileHelper::createDirectory(dirname($path));
-
-        $image = imagecreatetruecolor(100, 100);
-        imagejpeg($image, $path);
-
-        $file = File::create();
-        $file->loadDefaultValues();
-        $file->name = ucfirst($basename);
-        $file->basename = $basename;
-        $file->extension = 'jpg';
-        $file->width = 100;
-        $file->height = 100;
-        $file->size = filesize($path) ?: 0;
-        $file->populateFolderRelation($this->folder);
-
-        self::assertTrue($file->insert(), print_r($file->getErrors(), true));
-
-        return $file;
-    }
 }